+++ aspecd/analysis.py
@@ -284,6 +284,20 @@ class AnalysisStep(aspecd.utils.ToDictMixin):
 
         .. versionadded:: 0.5
 
+    dataset_type : :class:`str`
+        Full class name of the dataset that should be created
+
+        In case of returning a calculated dataset, packages derived from
+        the ASpecD framework may want to return their own instances of
+        :class:`aspecd.dataset.CalculatedDataset`.
+
+        Note that due to assigning some metadata, the class specified here
+        needs to conform to :class:`aspecd.dataset.CalculatedDataset`.
+
+        Default: "aspecd.dataset.CalculatedDataset"
+
+        .. versionadded:: 0.7
+
     description : :class:`str`
         Short description, to be set in class definition
 
@@ -312,6 +326,7 @@ class AnalysisStep(aspecd.utils.ToDictMixin):
         self.parameters = dict()
         self.result = None
         self.index = []
+        self.dataset_type = 'aspecd.dataset.CalculatedDataset'
         self.description = 'Abstract analysis step'
         self.comment = ''
         self.references = []
@@ -359,7 +374,7 @@ class AnalysisStep(aspecd.utils.ToDictMixin):
         .. versionadded:: 0.2
 
         """
-        dataset = aspecd.dataset.CalculatedDataset()
+        dataset = aspecd.utils.object_from_class_name(self.dataset_type)
         dataset.metadata.calculation.type = self.name
         dataset.metadata.calculation.parameters = self.parameters
         return dataset
@@ -383,6 +398,10 @@ class AnalysisStep(aspecd.utils.ToDictMixin):
         checks.
 
         """
+
+    def _assign_origdata_in_dataset(self):
+        if isinstance(self.result, aspecd.dataset.Dataset):
+            self.result._origdata = copy.deepcopy(self.result.data)  # noqa
 
     # noinspection PyUnusedLocal
     @staticmethod
@@ -472,7 +491,7 @@ class SingleAnalysisStep(AnalysisStep):
         """Perform the actual analysis step on the given dataset.
 
         If no dataset is provided at method call, but is set as property in
-        the SingleAnalysisStep object, the analysd method of the dataset
+        the SingleAnalysisStep object, the analyse method of the dataset
         will be called and thus the analysis added to the list of analyses
         of the dataset.
 
@@ -537,6 +556,7 @@ class SingleAnalysisStep(AnalysisStep):
             self._check_applicability()
             self._sanitise_parameters()
             self._perform_task()
+            self._assign_origdata_in_dataset()
 
     def _check_applicability(self):
         if not self.applicable(self.dataset):
@@ -653,6 +673,7 @@ class MultiAnalysisStep(AnalysisStep):
         self._check_applicability()
         self._sanitise_parameters()
         self._perform_task()
+        self._assign_origdata_in_dataset()
 
     def _check_applicability(self):
         for dataset in self.datasets:
@@ -757,6 +778,7 @@ class AggregatedAnalysisStep(AnalysisStep):
             self.result.data.data = \
                 np.reshape(self.result.data.data, (n_datasets, -1))
         self.result.data.axes[0].index = index
+        self._assign_origdata_in_dataset()
 
     def _check_and_prepare(self):
         if not self.datasets: