+++ aspecd/processing.py
@@ -1328,14 +1328,17 @@ class SliceExtraction(SingleProcessingStep):
 
 
     .. versionchanged:: 0.2
-       Parameter "index" renamed to "position" to reflect values to be
-       either indices or axis values
+        Parameter "index" renamed to "position" to reflect values to be
+        either indices or axis values
 
     .. versionadded:: 0.2
-       Slice positions can be given both, as axis indices and axis values
+        Slice positions can be given both, as axis indices and axis values
 
     .. versionadded:: 0.2
-       Works for ND datasets with N>1
+        Works for ND datasets with N>1
+
+    .. versionchanged:: 0.7
+        Sets dataset label to slice position (in axes units)
 
 
     Examples
@@ -1461,6 +1464,7 @@ class SliceExtraction(SingleProcessingStep):
             raise ValueError("Position out of axis range.")
 
     def _perform_task(self):
+        self.dataset.label = self._set_dataset_label()
         slice_object = self._get_slice()
         self.dataset.data.data = self.dataset.data.data[slice_object]
         for axis in self.parameters["axis"]:
@@ -1498,6 +1502,21 @@ class SliceExtraction(SingleProcessingStep):
     @staticmethod
     def _get_index(vector, value):
         return np.abs(vector - value).argmin()
+
+    def _set_dataset_label(self):
+        label = ''
+        for idx, axis in enumerate(self.parameters["axis"]):
+            if label:
+                label += ', '
+            if self.parameters["unit"] == "index":
+                value = self.dataset.data.axes[axis].values[
+                    self.parameters["position"][idx]]
+            else:
+                index = self._get_index(self.dataset.data.axes[axis].values,
+                                        self.parameters["position"][idx])
+                value = self.dataset.data.axes[axis].values[index]
+            label += "{} {}".format(value, self.dataset.data.axes[axis].unit)
+        return label
 
 
 class RangeExtraction(SingleProcessingStep):
@@ -1987,6 +2006,10 @@ class Averaging(SingleProcessingStep):
         Raised if axis is out of bounds for given dataset
 
 
+    .. versionchanged:: 0.7
+        Sets dataset label to averaging range (in axes units)
+
+
     Examples
     --------
     For convenience, a series of examples in recipe style (for details of
@@ -2079,6 +2102,7 @@ class Averaging(SingleProcessingStep):
             raise ValueError("Given range out of axis range.")
 
     def _perform_task(self):
+        self.dataset.label = self._set_dataset_label()
         range_ = self._get_range()
         if self.parameters["axis"] == 0:
             self.dataset.data.data = \
@@ -2114,7 +2138,7 @@ class Averaging(SingleProcessingStep):
 
     def _get_range(self):
         if self.parameters["unit"] == "index":
-            range_ = self.parameters["range"]
+            range_ = copy.copy(self.parameters["range"])
             if range_[1] > 0:
                 range_[1] += 1
         else:
@@ -2130,6 +2154,14 @@ class Averaging(SingleProcessingStep):
     @staticmethod
     def _get_index(vector, value):
         return np.abs(vector - value).argmin()
+
+    def _set_dataset_label(self):
+        range_ = self._get_range()
+        range_[1] -= 1
+        values = self.dataset.data.axes[self.parameters["axis"]].values[range_]
+        unit = self.dataset.data.axes[self.parameters["axis"]].unit
+        label = '{}-{} {}'.format(values[0], values[1], unit)
+        return label
 
 
 class ScalarAxisAlgebra(SingleProcessingStep):