+++ docs/changelog.rst
@@ -5,6 +5,40 @@ Changelog
 This page contains a summary of changes between the official ASpecD releases. Only the biggest changes are listed here. A complete and detailed log of all changes is available through the `GitHub Repository Browser <https://github.com/tillbiskup/aspecd/commits/master>`_.
 
 
+Version 0.7.0
+=============
+
+Released 2022-01-30
+
+
+New features
+------------
+
+* :class:`aspecd.tasks.FigurereportTask` for creating figure captions that can, *e.g.*, be included in other documents
+* Attributes ``labelspacing`` and ``fontsize`` in :class:`aspecd.plotting.LegendProperties`
+* Attribute ``output`` in :class:`aspecd.tasks.ModelTask` controlling the type of output returned (dataset or model)
+* Method :meth:`aspecd.model.Model.evaluate` for fast evaluation of models without any checks (useful in context of fitting)
+* Attribute ``dataset_type`` in :class:`aspecd.analysis.AnalysisStep` to define type of calculated dataset that gets returned
+* :class:`aspecd.plotting.MultiPlotter1D` and :class:`aspecd.plotting.MultiPlotter1DStacked` with parameter "tight" for tight axes and "switch_axes" for switching axes
+* :class:`aspecd.plotting.SinglePlotter1D` with parameter "switch_axes" for switching axes
+* :class:`aspecd.plotting.AxesProperties`: angles of the axes tick labels can be set using the ``xticklabelangle`` and ``yticklabelangle`` properties
+
+
+Changes
+-------
+
+* :class:`aspecd.processing.SliceExtraction` sets dataset label to slice position
+* :class:`aspecd.processing.Averaging` sets dataset label to averaging range
+
+
+Fixes
+-----
+
+* Dataset importer does not override dataset label.
+* AnalysisSteps assign data to _origdata attribute if result is dataset
+* MultiprocessingTask correctly sets label of resulting datasets
+
+
 Version 0.6.4
 =============